Add remote setting for u-blox lea-m8f config
Ground truth coordinates can be programmed to GPS modules from gui.
Decreases setup effort if ground truth coordinates are needed for
precise synchronization.
A file for generating and sending the bitstream to the gps modules is
added.
The default behaviour of the "set position" button for the receivers is
changed: Dialog for setting the GPS coordinates appears. Manual
estimates can still be set from map by pressing map estimate.
RPC requests are added to set the coordinates for each receiver.