Implement manual offset compensation (ftp delays in ns)
The phased lock loop in the USRPs causes discrete and fixed offsets.
Since no elegant solution for this problem could be found, a
manual offset compensation is necessary.
All delay offsets are integer multiples of 5ns, consequently delays
will be stored and processed in nanoseconds from now on.
The offset that was determined by the user can be set for each receiver
in the table model in tab 4.