Fix handling of missing samples
Reset ONLY those receivers that yield "delayed" samples
compared to the latest timestamp. However, missing sets of samples
remain problematic if a fixed number of acquisitions should be taken.
Consequently, remove this in all cases except single acquisitions.
It is replaced by starting an infinite reception terminated
by a stop-loop-like statement. After that, the fusion center
goes into a "flush" state in which the received samples are not
processed.